refactor: codingzone 도메인 관련 infra 파일 이동
